BNL_T03 updates to address bias drift issues + updates to metrology scripts + addition of gain stability acq functionality to T08 eo_acq script

+++ BNL_T03/flat_acq/v0/ccseoflat-twosampleflux.py
@@ -94,7 +94,7 @@ try:
     print "Now collect some parameters from the config file"
     lo_lim = float(eolib.getCfgVal(acqcfgfile, 'FLAT_LOLIM', default='0.1'))
     hi_lim = float(eolib.getCfgVal(acqcfgfile, 'FLAT_HILIM', default='120.0'))
-    bcount = int(eolib.getCfgVal(acqcfgfile, 'FLAT_BCOUNT', default = "2"))
+    bcount = float(eolib.getCfgVal(acqcfgfile, 'FLAT_BCOUNT', default = "2"))
     wl     = float(eolib.getCfgVal(acqcfgfile, 'FLAT_WL', default = "550.0"))
     imcount = 2
